+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Important note - with version 3.0.0 accesory moved to Power Meter custom Characteristic, due to Apple HomeKit limitations right now only in EVE app displayed correctly, in HomeKit displayed as 'Unsupported'. If U want to use old CO2 sensor style just still with 2.x.x version
 
+## [3.4.0] - (29.01.2021)
+## Changs
+- added check for installed Q-Relay 
+- added check for installed Encharge 
+- added check for installed Meters
+- reconfigured config menu
+- code rebuild
+
 ## [3.3.15] - (26.01.2021)
 ## Changs
 - power Net and Total Max fixes